% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTree.R
\name{PlotTree}
\alias{PlotTree}
\title{Plotting an option tree}
\usage{
PlotTree(x)
}
\arguments{
\item{x}{A matrix of option prices as returned by Derivative_Security.}
}
\value{
Plots the option tree in the viewer plane..
}
\description{
Description
}
\examples{
PlotTree(Derivative_Security(S, K, N, r, d, u, type));
}
