% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Derivative_Security.R
\name{Derivative_Security}
\alias{Derivative_Security}
\title{Derivative Security}
\usage{
Derivative_Security(S, K, N, r, d, u, type)
}
\arguments{
\item{S}{Description of parameter S.}

\item{K}{Description of parameter K.}

\item{N}{Description of parameter N.}

\item{r}{Description of parameter r.}

\item{d}{Description of parameter d.}

\item{u}{Description of parameter u.}

\item{type}{Description of parameter type.}
}
\value{
What it returns.
}
\description{
Description
}
\examples{
Derivative_Security(S = 4, K = 4, N = 4, r = 0.25, d = 0.5, u = 2, type = "European_put")
}
