% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTree.R
\name{PlotTree}
\alias{PlotTree}
\title{Plotting an option tree}
\usage{
PlotTree(x)
}
\arguments{
\item{x}{A matrix of either option prices or deltas as returned by Derivative_Security().}
}
\value{
Plots the option tree in the viewer plane..
}
\description{
Description
}
\examples{
option <- Derivative_Security(S = 4, K = 4, N = 4, r = 0.25, d = 0.5, u = 2, type = "European_put")
PlotTree(option$derivative_prices)
PlotTree(option$delta)
}
